DABOM-302 refactor: 스타일가이드 준수 개선
- Response DTO를 dto/response/ 패키지로 이동 (패키지 구조 규칙)
- NotificationSlice 도메인 모델 생성 — Service가 Entity 기반 결과 반환
- NotificationService: Response DTO 직접 반환 제거
  getNotifications() → NotificationSlice, getUnreadCount() → long
- NotificationController: DTO 변환 책임을 Controller로 이동
- WebPushController: unsubscribe() ApiResponse<Void> 반환으로 변경
- WebPushServiceImpl: sendToUser/sendToFamily에 @Transactional(readOnly=true) 추가
- NotificationLogRepositoryImpl: 매직넘버 30 → RETENTION_DAYS 상수 추출
