+++ src/main/java/com/project/domain/notification/dto/response/NotificationResponse.java
@@ -1,4 +1,4 @@
-package com.project.domain.notification.dto;
+package com.project.domain.notification.dto.response;
 
 import java.time.LocalDateTime;
