+++ src/main/java/com/project/domain/notification/dto/response/UnreadCountResponse.java
@@ -1,3 +1,3 @@
-package com.project.domain.notification.dto;
+package com.project.domain.notification.dto.response;
 
 public record UnreadCountResponse(long unreadCount) {}
