+++ src/main/java/com/project/domain/webpush/service/WebPushServiceImpl.java
@@ -97,6 +97,7 @@ public class WebPushServiceImpl implements WebPushService {
         pushSubscriptionRepository.delete(subscription);
     }
 
+    @Transactional(readOnly = true)
     @Override
     public void sendToUser(Long customerId, String title, String message) {
         PushSubscription subscription =
@@ -110,6 +111,7 @@ public class WebPushServiceImpl implements WebPushService {
         sendPushNotification(subscription, title, message);
     }
 
+    @Transactional(readOnly = true)
     @Override
     public void sendToFamily(Long familyId, String title, String message) {
         List<Long> customerIds = familyMemberRepository.findCustomerIdsByFamilyId(familyId);
